[MCJ-1557] refactor: 환불 요청 목록 DTO 매핑을 from 팩토리로 분리
